# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Battery Data Intelligence Engine",
    "description": "Intelligence API backing the admin dashboard. Protected routes verify Clerk session JWTs via JWKS; see app/api/deps.py.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/materials": {
      "get": {
        "tags": [
          "materials"
        ],
        "summary": "List Materials",
        "operationId": "list_materials_api_v1_materials_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "is_ira_critical",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Ira Critical"
            }
          },
          {
            "name": "is_eu_crma_critical",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Eu Crma Critical"
            }
          },
          {
            "name": "has_mismatched_mappings",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Has Mismatched Mappings"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_MaterialListItem_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/materials/{material_id}": {
      "get": {
        "tags": [
          "materials"
        ],
        "summary": "Get Material",
        "operationId": "get_material_api_v1_materials__material_id__get",
        "parameters": [
          {
            "name": "material_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Material Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/materials/{material_id}/notes": {
      "get": {
        "tags": [
          "materials"
        ],
        "summary": "List Material Notes",
        "operationId": "list_material_notes_api_v1_materials__material_id__notes_get",
        "parameters": [
          {
            "name": "material_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Material Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnalystNoteRead"
                  },
                  "title": "Response List Material Notes Api V1 Materials  Material Id  Notes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "materials"
        ],
        "summary": "Create Material Note",
        "operationId": "create_material_note_api_v1_materials__material_id__notes_post",
        "parameters": [
          {
            "name": "material_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Material Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalystNoteCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalystNoteRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/materials/{material_id}/verified": {
      "patch": {
        "tags": [
          "materials"
        ],
        "summary": "Set Material Verified",
        "operationId": "set_material_verified_api_v1_materials__material_id__verified_patch",
        "parameters": [
          {
            "name": "material_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Material Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifiedUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifiedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalystNoteCreate": {
        "properties": {
          "note_type": {
            "type": "string",
            "enum": [
              "data_error",
              "missing_data",
              "outdated",
              "other"
            ],
            "title": "Note Type"
          },
          "note_text": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 3,
            "title": "Note Text"
          }
        },
        "type": "object",
        "required": [
          "note_type",
          "note_text"
        ],
        "title": "AnalystNoteCreate",
        "description": "Request body for per-entity POST \u2026/notes endpoints.\n\nClients send only ``note_type`` and ``note_text``.  The route handler\ninjects ``entity_type`` and ``entity_id`` from the URL path before writing\nto ``analyst_notes``."
      },
      "AnalystNoteRead": {
        "properties": {
          "note_type": {
            "type": "string",
            "enum": [
              "data_error",
              "missing_data",
              "outdated",
              "other"
            ],
            "title": "Note Type"
          },
          "note_text": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 3,
            "title": "Note Text"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "entity_type": {
            "type": "string",
            "enum": [
              "company",
              "company_material_exposure",
              "company_supply_relationship",
              "company_regulation_exposure",
              "company_vehicle_model",
              "material",
              "hs_code_material_mappings",
              "regulation",
              "risk_event",
              "facility",
              "battery_chemistry"
            ],
            "title": "Entity Type"
          },
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "note_type",
          "note_text",
          "id",
          "entity_type",
          "entity_id",
          "created_at",
          "updated_at"
        ],
        "title": "AnalystNoteRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MaterialDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "canonical_name": {
            "type": "string",
            "title": "Canonical Name"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "symbol_or_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol Or Code"
          },
          "hs_codes": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Hs Codes"
          },
          "criticality_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Criticality Score"
          },
          "primary_producing_countries": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Primary Producing Countries"
          },
          "price_unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Unit"
          },
          "is_ira_critical_mineral": {
            "type": "boolean",
            "title": "Is Ira Critical Mineral"
          },
          "is_eu_crma_critical": {
            "type": "boolean",
            "title": "Is Eu Crma Critical"
          },
          "patent_occurrence_trend": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patent Occurrence Trend"
          },
          "data_availability": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Availability"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "criticality_signals": {
            "items": {
              "$ref": "#/components/schemas/CriticalitySignalRead"
            },
            "type": "array",
            "title": "Criticality Signals",
            "default": []
          },
          "chemistry_uses": {
            "items": {
              "$ref": "#/components/schemas/ChemistryUseRead"
            },
            "type": "array",
            "title": "Chemistry Uses",
            "default": []
          },
          "hs_mappings": {
            "items": {
              "$ref": "#/components/schemas/HsMappingRead"
            },
            "type": "array",
            "title": "Hs Mappings",
            "default": []
          },
          "mapping_health": {
            "$ref": "#/components/schemas/MappingHealth"
          }
        },
        "type": "object",
        "required": [
          "id",
          "canonical_name",
          "is_ira_critical_mineral",
          "is_eu_crma_critical",
          "created_at",
          "updated_at"
        ],
        "title": "MaterialDetail"
      },
...
      "PaginatedResponse_MaterialListItem_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/MaterialListItem"
            },
            "type": "array",
            "title": "Data"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "data",
          "total",
          "page",
          "limit"
        ],
        "title": "PaginatedResponse[MaterialListItem]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VerifiedResponse": {
        "properties": {
          "verified": {
            "type": "boolean",
            "title": "Verified"
          }
        },
        "type": "object",
        "required": [
          "verified"
        ],
        "title": "VerifiedResponse",
        "description": "Minimal response returned by all PATCH \u2026/verified endpoints."
      },
      "VerifiedUpdate": {
        "properties": {
          "verified": {
            "type": "boolean",
            "title": "Verified"
          }
        },
        "type": "object",
        "required": [
          "verified"
        ],
        "title": "VerifiedUpdate",
        "description": "Request body for PATCH \u2026/verified endpoints."
      }
    }
  }
}